-Move all files to source dir, now build and tests should work -Move all tests to typescript -Move all tests from equal to strictEqual since equal is deprecated
