+++ azure-kusto-ingest/package.json
@@ -32,11 +32,14 @@
     "azure-kusto-data": "latest",
     "moment": "^2.22.2",
     "request": "^2.88.0",
+    "source-map-support": "^0.5.19",
     "uuid": "^3.3.2",
     "uuid-validate": "0.0.3"
   },
   "devDependencies": {
+    "@types/mocha": "^8.2.0",
     "@types/node": "^14.14.13",
+    "@types/sinon": "^9.0.9",
     "@types/uuid": "^8.3.0",
     "@types/uuid-validate": "0.0.1",
     "eslint": "^6.8.0",
