+++ azure-kusto-ingest/source/descriptors.ts
