+++ azure-kusto-ingest/source/status.ts
