+++ azure-kusto-ingest/source/statusQ.ts
